requirements.txt
fastapi
uvicorn
pydantic
joblib==1.3.2
pandas>=2.1.3
scikit-learn>=1.3.2
httpx==0.25.1
python-dotenv==1.0.0
numpy>=1.26.0
redis==5.0.1
python-json-logger==2.0.7
slowapi==0.1.9
ortools
